Rename CharcoalPopupProtocol
